feat: moving the getProductCost method from PriceView to make it reusable
